requirements.txt
absl-py==1.0.0
ai-benchmark==0.1.2
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
black==22.3.0
bleach==4.1.0
cachetools==5.0.0
certifi==2021.10.8
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
commonmark==0.9.1
coverage==6.3.2
dacite==1.6.0
debugpy==1.6.0
decorator==5.1.1
distlib==0.3.4
Django==4.0.3
django-crispy-forms==1.14.0
dnspython==2.1.0
docutils==0.18.1
entrypoints==0.4
executing==0.8.3
filelock==3.6.0
flake8==4.0.1
flatbuffers==2.0
gast==0.5.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
huggingface-hub==0.4.0
identify==2.4.12
idna==2.7
imap-tools==0.50.2
importlib-metadata==4.11.3
iniconfig==1.1.1
ipykernel==6.11.0
ipython==8.2.0
isort==5.10.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.2.1
jupyter-core==4.9.2
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
libclang==13.0.0
Markdown==3.3.6
matplotlib-inline==0.1.3
mccabe==0.6.1
mypy-extensions==0.4.3
nest-asyncio==1.5.5
nodeenv==1.6.0
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
parallelformers==1.2.4
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.8.2
platformdirs==2.5.1
pluggy==1.0.0
pre-commit==2.18.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
pure-eval==0.2.2
py==1.11.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.11.2
pymongo==4.0.1
pynvml==11.4.1
pyparsing==3.0.7
pytest==7.1.1
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pywin32==303
pywin32-ctypes==0.2.0
PyYAML==6.0
pyzmq==22.3.0
readme-renderer==34.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.8.0
rfc3986==2.0.0
rich==12.1.0
rsa==4.8
sacremoses==0.0.49
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
tokenizers==0.11.6
toml==0.10.2
tomli==2.0.1
torch==1.11.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
transformers==4.17.0
twine==4.0.0
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0
